refactor(shared): extract the blob-download anchor block into a helper
The "create an anchor, point it at an object URL, click it, revoke" block
was written out three times, and the copies had already drifted: the two
CSV export sites attached the anchor to document.body before clicking and
removed it after, the binary download in representation.service did not.

triggerBlobDownload lands in vre/shared/app-common — the only shared lib
both consuming libs already depend on, so no new dependency edge — and
keeps the append/remove behaviour as the single implementation.

Encoding stays at the call sites: which MIME type to declare and whether
the content needs a UTF-8 BOM are per-format decisions, not download
mechanics. representation.service is unaffected by that concern; it
streams res.body straight from the server and never builds a Blob.

DEV-6993
