+++ libs/vre/shared/app-common/src/index.ts
@@ -1,5 +1,6 @@
 export * from './lib/resource.service';
 export * from './lib/custom-regex';
+export * from './lib/download-file';
 export * from './lib/dsp-resource';
 export * from './lib/property-info-values.interface';
 export * from './lib/generateProperty';
